feat: check doctor availability and get appointments

+++ be/internal/domain/entity/bill.go
@@ -9,7 +9,7 @@ type Bill struct {
 	CashingOfficerId *int64
 	MedicalRecordId  int64
 	Description      *string
-	TotalPrice       int
+	TotalPrice       float32
 	Status           string `gorm:"type:bill_status_slug;default:'unpaid'"`
 	CreatedAt        time.Time
 	UpdatedAt        time.Time
